Improve commenting in LogicSimulator.mlapp
Improved the commenting in LogicSimulator.mlapp to make the code easier to follow.

Added a 'new' option to the menubar under 'File' that simply wipes the circuit clean.

NOTE: a TODO comment was removed regarding not being able to click on points in the circuit plot occupied by drawn lines. This is not yet fixed, the issue note was just moved to an out-of-source issue tracker.